Added more client command, also added CsvHelper library so the functions that return lists of objects can be converted to actual objects

+++ SoftEtherVPNCmdNET/VpnCmd.cs
@@ -13,6 +13,7 @@ namespace SoftEtherVPNCmdNET
         protected string password;
         protected Process app;
         protected ProcessStartInfo processStartInfo;
+        protected bool timeoutOverride { get; set; }
         private int timeout;
         private bool client;
         public VpnCmd(string hostport, string pass, bool client, int timeout)
@@ -21,11 +22,11 @@ namespace SoftEtherVPNCmdNET
             this.password = pass;
             this.client = client;
             this.timeout = timeout;
+            this.timeoutOverride = false;
 
             processStartInfo = new ProcessStartInfo();
 
             processStartInfo.FileName = "vpncmd";
-
             processStartInfo.UseShellExecute = false;
             processStartInfo.RedirectStandardOutput = true;
             processStartInfo.WindowStyle = ProcessWindowStyle.Hidden;
@@ -38,8 +39,12 @@ namespace SoftEtherVPNCmdNET
             app.StartInfo.Arguments = hostport + (client == true?" /CLIENT ":" /SERVER ") + (password == "" ? " " : " /PASSWORD " + password + " ") + " /CSV /CMD " + command + " " + arguments;
             app.Start();
 
-            //kills the process if it hangs more than specified timeout period
-            Timer t = new Timer(f => { app.Kill(); },null,this.timeout,Timeout.Infinite);
+            //override the timeout if needed so the process wont be killed
+            if (!timeoutOverride) 
+            {
+                //kills the process if it hangs more than specified timeout period
+                Timer t = new Timer(f => { app.Kill(); }, null, this.timeout, Timeout.Infinite);
+            }
             
             string output = app.StandardOutput.ReadToEnd();
             var words = output.Split(' ', '\n', '\r', '\t');
@@ -47,8 +52,15 @@ namespace SoftEtherVPNCmdNET
 
             if (matchquery.Count() >= 1)
                 return new Response(false, output);
-
-            app.WaitForExit(this.timeout);
+            if (!timeoutOverride) //override the timeout if needed
+            {
+                app.WaitForExit(this.timeout);
+                timeoutOverride = false;
+            }
+            else
+            {
+                app.WaitForExit();
+            }
             return new Response(true, output);
         }
     }
